[Company Employee][Hans] feature: Change the type of salery in V1.1__create_employee.sql

+++ src/main/resources/db/migration/V1.1__create_employee.sql
@@ -3,7 +3,7 @@ create table employee(
     name VARCHAR(255),
     gender VARCHAR(255),
     age INT,
-    salery INT,
+    salery DOUBLE,
     company_id INT,
     foreign key (company_id) references company(company_id)
 )
